* 64-bit configure support for Linux x86_64  * fixed a stupid bug in generating parse exceptions (how did that ever work before???)  * fixed a deadlock in adding new (unknown) character encodings in include/qore/charset.h  * fixed a bug in the QoreString::vsprintf() design and usage (retries when the buffer was too small would cause a segfault)  * set library and module dir to use lib64 for 64-bit builds on Linux
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@63 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/ql_debug.cc
@@ -46,7 +46,7 @@ class Object *CallStack::getPrevStackObject()
 
 static class QoreNode *f_dbg_get_object_ptr(class QoreNode *params, ExceptionSink *xsink)
 {
-   return new QoreNode((int64)(unsigned)thread_list[gettid()].callStack->getPrevStackObject());
+   return new QoreNode((int64)(unsigned long)thread_list[gettid()].callStack->getPrevStackObject());
 }
 
 //static
